2023-spring/cs-281/sample_python_code/requirements.txt
psycopg2==2.9.6
python-dotenv==1.0.0
